fix(words): a floating VML shape paints on the layer its z-index asks for
A `v:shape` carries a CSS `z-index` and nothing read it, so every floating VML
shape was painted in document order in front of the text. Both halves of that
are wrong, and neither presents as a z-order fault.

Forward, it presents as missing content. `JEMIT_Template.docx` puts its journal
title in a `v:shape` declaring `z-index:251659264` and its grey masthead band in
an ordinary `wp:inline` picture declared in the same paragraph just after it, so
the picture painted over the title and the band came out empty. The words were
in the PDF's text layer the whole time — `pdftotext` reads all eight of them off
our own page 1 — which is exactly the class of defect no gate column can see.
Backward, a negative `z-index` is how Word writes a watermark, and drawing that
over the text rather than under it is the same fault mirrored.

LibreOffice keeps the declaration verbatim through the VML importer
(`oox/source/vml/vmlshapecontext.cxx:536`, handed on as `VML-Z-ORDER` at
`oox/source/vml/vmlshape.cxx:408`) and the mapper then sets the shape opaque
exactly when it is not negative — `setPropertyValue(PROP_OPAQUE, Any(zOrder >=
0))`, `sw/source/writerfilter/dmapper/DomainMapper_Impl.cxx:5157` and `:5203`.
So the sign is the layer and the magnitude is the order within it.

The second rule is that a shape declaring a `z-index` outranks *every*
DrawingML `relativeHeight`, whatever the two numbers are.
`GraphicZOrderHelper::adjustRelativeHeight`
(`sw/source/writerfilter/dmapper/GraphicHelpers.cxx:279-330`) says so in as many
words — "in general, all z-index-defined shapes appear on top of relativeHeight
graphics regardless of the value" — and implements it by pushing every
DrawingML anchor below zero (`GraphicImport.cxx:695`). It matters because the
two families sit in the same numeric neighbourhood: `JEMIT`'s text box declares
`z-index:251659264` and the corpus's `wp:anchor`s declare `relativeHeight`
around 251 660 000, so a reader comparing them directly interleaves them. Here
the separation is done the other way about, by lifting a `z-index` by 2^32 —
one past the top of `ST_RelativeHeight`'s unsigned range — which leaves every
stored DrawingML order exactly as it was. `PageFrame.ZOrder` widens from `uint`
to `long` to hold it.

Measured against LibreOffice 26.2.4.2 (mean per-page ink; `soffice` here is
24.2.7.2 and is not the target). Rendering the 158 DOCX documents that carry a
`z-index` or a `w:hideMark` before and after, **16 of them changed bytes for
this rule and no page count moved anywhere**. Twelve are ink-neutral: the paint
order changed and the pixels did not. The four that moved:

| document | before | after |
|---|---:|---:|
| `067_Work_Breakdown_Structure_Template_Gray_Theme` | 10.83 | 10.45 |
| `069_Work_Breakdown_Structure_Template_Professional_Format` | 13.31 | 13.18 |
| `065_Work_Breakdown_Structure_Template_Blue_Theme` | 5.81 | **5.87** |
| `JEMIT_Template.docx` | 15.61 | **15.72** |

**Two of those are worse and one of them is the document this was found on**,
so it is worth saying plainly what happened. `JEMIT`'s masthead title is now
drawn, correctly, where it was buried before — but our grey masthead band sits
about 12 pt below the reference's and is taller, so the newly drawn title lands
12 pt low as well. Two mismatched blobs score worse than one missing one. The
band's offset is a separate defect and is untouched here; page 1's ink went
21.89 to 22.32 and the page is visibly closer to the reference than it was.
`065` moves by 0.06, which is one shape's worth of overlap changing hands.

Every non-fidelity test project run: 5459 passed, 0 failed. `Paperless.Fidelity`
reports 15 failed of 552, and the same 13 of those covered by the four relevant
test classes fail identically on the unmodified tree — they are the documented
24.2.7.2-against-26.2.4.2 gap, not this change.

+++ dotnet/src/Paperless.WordProcessing/Layout/PageFrame.cs
@@ -394,8 +394,19 @@ public sealed record PageFrame
     /// Zero when the anchor does not declare one, which sorts it below anything that does — and since
     /// the sort is stable, equal values keep document order, which is Word's own tie-break.
     /// </para>
+    /// <para>
+    /// <strong>It is a <c>long</c> because two different declarations land in it and they do not share
+    /// a range.</strong> DrawingML's <c>relativeHeight</c> is an unsigned 32-bit value; VML's
+    /// <c>z-index</c> is signed, and LibreOffice sorts <em>every</em> shape that declares one above
+    /// <em>every</em> <c>relativeHeight</c> whatever the two numbers are —
+    /// <c>GraphicZOrderHelper::adjustRelativeHeight</c>, <c>sw/source/writerfilter/dmapper/
+    /// GraphicHelpers.cxx:279-330</c>: "in general, all z-index-defined shapes appear on top of
+    /// relativeHeight graphics regardless of the value". <see cref="Ooxml.DocxVmlFrames"/> therefore
+    /// offsets a <c>z-index</c> by 2^32, which is above the whole unsigned range and keeps both
+    /// families' internal ordering intact.
+    /// </para>
     /// </remarks>
-    public uint ZOrder { get; init; }
+    public long ZOrder { get; init; }
 
     /// <summary>
     /// The shape's <c>a:prstGeom/@prst</c>, or null when it states none and is a plain box.
